requirements.txt
faiss-cpu==1.13.2
networkx==3.4.2
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.3.0
torch==2.7.1+cu118
torch-geometric==2.7.0
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
transformers==5.3.0
